Abstract filter functions.
- Introduce AttributeFilterData and SymbolFilterData that store information about a symbol necessary for filtering.
- Make fiter rules work with these two classes.
- Move Roslyn-specific code to RoslynFilterData.
- Move filter config file loading code into ConfigFilterRule.
