+++ src/Microsoft.DocAsCode.Metadata.ManagedReference.Common/Filters/AttributeFilterInfo.cs
@@ -7,7 +7,6 @@ namespace Microsoft.DocAsCode.Metadata.ManagedReference
     using System.Collections.Generic;
     using System.Linq;
 
-    using Microsoft.CodeAnalysis;
     using YamlDotNet.Serialization;
 
     [Serializable]
@@ -22,28 +21,25 @@ namespace Microsoft.DocAsCode.Metadata.ManagedReference
         [YamlMember(Alias = "ctorNamedArguments")]
         public Dictionary<string, string> ConstructorNamedArguments { get; set; } = new Dictionary<string, string>();
 
-        public bool ContainedIn(ISymbol symbol)
+        public bool ContainedIn(SymbolFilterData symbol)
         {
             bool result = false;
-            var attributes = symbol.GetAttributes();
+            var attributes = symbol.Attributes;
             foreach (var attribute in attributes)
             {
-                INamedTypeSymbol attributeClass = attribute.AttributeClass;
-                if (Uid != VisitorHelper.GetId(attributeClass))
+                if (Uid != attribute.Id)
                 {
                     continue;
                 }
 
                 // arguments need to be a total match of the config
-                IEnumerable<string> arguments = attribute.ConstructorArguments.Select(arg => GetLiteralString(arg));
-                if (!ConstructorArguments.SequenceEqual(arguments))
+                if (!ConstructorArguments.SequenceEqual(attribute.ConstructorArguments))
                 {
                     continue;
                 }
 
                 // namedarguments need to be a superset of the config
-                Dictionary<string, string> namedArguments = attribute.NamedArguments.ToDictionary(pair => pair.Key, pair => GetLiteralString(pair.Value));
-                if (!ConstructorNamedArguments.Except(namedArguments).Any())
+                if (!ConstructorNamedArguments.Except(attribute.ConstructorNamedArguments).Any())
                 {
                     result = true;
                     break;
@@ -51,29 +47,6 @@ namespace Microsoft.DocAsCode.Metadata.ManagedReference
             }
 
             return result;
-        }
-
-        private static string GetLiteralString(TypedConstant constant)
-        {
-            var type = constant.Type;
-            var value = constant.Value;
-
-            if (type.TypeKind == TypeKind.Enum)
-            {
-                var namedType = (INamedTypeSymbol)type;
-                var pairs = (from member in namedType.GetMembers().OfType<IFieldSymbol>()
-                             where member.IsConst && member.HasConstantValue
-                             select Tuple.Create(member.Name, member.ConstantValue)).ToDictionary(tuple => tuple.Item2, tuple => tuple.Item1);
-
-                return $"{VisitorHelper.GetId(namedType)}.{pairs[value]}";
-            }
-
-            if (value is ITypeSymbol)
-            {
-                return VisitorHelper.GetId((ITypeSymbol)value);
-            }
-
-            return value.ToString();
         }
     }
 }
